chat_api/pyproject.toml
[project]
name = "chatbot-api"
version = "0.2.0"
description = "API Chatbot  for RMUTL NAN"
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.15.1",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.17",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.2",
    "uvicorn>=0.34.2",
    "bcrypt>=4.3.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
]
